+++ test/test.cc
@@ -341,6 +341,12 @@ boost::property_tree::ptree make_config(const std::string& path_prefix,
           "max_matrix_distance": 200000.0,
           "max_matrix_location_pairs": 2500
         },
+        "auto_pedestrian": {
+          "max_distance": 500000.0,
+          "max_locations": 50,
+          "max_matrix_distance": 200000.0,
+          "max_matrix_location_pairs": 2500
+        },
         "bus": {
           "max_distance": 5000000.0,
           "max_locations": 50,
@@ -414,8 +420,8 @@ boost::property_tree::ptree make_config(const std::string& path_prefix,
           "max_locations": 50,
           "max_matrix_distance": 200000.0,
           "max_matrix_location_pairs": 2500,
-          "max_transit_walking_distance": 10000,
-          "min_transit_walking_distance": 1
+          "max_multimodal_walking_distance": 10000,
+          "min_multimodal_walking_distance": 1
         },
         "skadi": {
           "max_shape": 750000,